+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Features
 
  - Add ability to store multiple measurements for Outputs ([#562](https://github.com/kizniche/mycodo/issues/562))
+ - Add Calibration: Atlas Scientific EZO Pump ([#562](https://github.com/kizniche/mycodo/issues/562))
  - Add ability to enable Daemon debug mode from Configuration page